third_party/WebKit/LayoutTests/http/tests/xmlhttprequest/response-blob-size.html

<html>
<body>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
var t = async_test("Test Blob.size of response blob received for XMLHttpRequest");
t.step(function() {
    var req = new XMLHttpRequest;
    var i = 0;
    // Run the test multiple times using the same XMLHttpRequest instance to
    // check if all states are cleared correctly.
    var doRequest = function() {
        req.open('GET', '../resources/test.ogv', true);
        req.responseType = 'blob';
        req.onreadystatechange = t.step_func(function() {
            if (req.readyState == req.DONE) {
                assert_equals(req.status, 200, "Status");
                assert_equals(req.response.size, 103746, "Size");
                if (++i < 3)
                    doRequest();
                else
                    t.done();
            }
        });
        req.send(null);
    };
    doRequest();
});
</script>
</body>
